pyproject.toml
[project]
name = "deeper-level"
version = "0.1.0"
description = "World Labs API example for generating 3D worlds"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
generate-world = "generate_world:main"
